refactor(accessToken): add access_token requests to dedicated file
